Fix Box, Flex bug
Using `system-components` to add "height" to Flex and Box introduced a
bug when including this lib in CRA. All of the Box / Flex attributes not
defined in this repo's definitions were missing, and passed directly to
the dom.

This was fixed by adding the height helper directly via
`styled-components` and `styled-system`.